protect contact operations from other users

+++ controllers/contactController.js
@@ -27,6 +27,7 @@ const createContact = asyncHandler(async (req,res)=>{
             name,
             email,
             phone,
+            user_id: req.user.id,
         }
     );
     res.status(201).json({contact});
@@ -54,6 +55,10 @@ const deleteContact = asyncHandler(async (req,res)=>{
         res.status(404);
         throw new Error ("contact not found");
     }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error ("You cant delete another users contact, permission denied");
+    }
     await Contact.findByIdAndRemove(req.params.id);
      console.log("contact.remove");
     res.status(200).json(contact);
@@ -68,6 +73,10 @@ const updateContact = asyncHandler(async (req,res)=>{
     if (!contact){
         res.status(404);
         throw new Error ("contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error ("You cant update another users contact, permission denied");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
